# packages/decent-offchain/src/api/static/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Decent API",
    "description": "API for the Decent protocol",
    "version": "0.1.0",
    "contact": {
      "name": "Decent DAO",
      "url": "https://decentdao.org"
    }
  },
  "servers": [
    {
      "url": "https://decent-offchain.up.railway.app",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Localhost"
    }
  ],
  "paths": {
...
    "/auth/verify": {
      "post": {
        "tags": ["auth"],
        "summary": "Verify a SIWE message and signature",
        "description": "Verify a Sign-In with Ethereum message and signature",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string"
                  },
                  "signature": {
                    "type": "string"
                  }
                },
                "required": ["message", "signature"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/User"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["auth"],
        "summary": "Get current user information",
        "description": "Retrieve information about the currently authenticated user",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/User"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["auth"],
        "summary": "Log out current user",
        "description": "Log out the currently authenticated user",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Logout"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/d/{chainId}/{address}/proposals/{slug}/comments": {
      "get": {
        "tags": ["comments"],
        "summary": "Get proposal comments",
        "description": "Retrieve all comments for a specific proposal",
        "security": [],
        "parameters": [
          {
            "name": "chainId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Comment"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["comments"],
        "summary": "Create comment",
        "description": "Create a new comment on a proposal",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "chainId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewComment"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Comment"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/d/{chainId}/{address}/proposals/{slug}/comments/{id}": {
      "put": {
        "tags": ["comments"],
        "summary": "Update comment",
        "description": "Update an existing comment",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "chainId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewComment"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Comment"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["comments"],
        "summary": "Delete comment",
        "description": "Delete an existing comment",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "chainId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Comment"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {},
          "error": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "nullable": true
              },
              "message": {
                "type": "string"
              },
              "code": {
                "type": "string",
                "nullable": true
              },
              "details": {
                "nullable": true
              }
            },
            "required": ["message"]
          }
        },
        "required": ["success"]
      },
...
      "Permissions": {
        "type": "object",
        "properties": {
          "isProposer": {
            "type": "boolean"
          },
          "isVoter": {
            "type": "boolean"
          },
          "isSigner": {
            "type": "boolean"
          },
          "isModerator": {
            "type": "boolean"
          }
        },
        "required": ["isProposer", "isVoter", "isSigner", "isModerator"]
      },
      "User": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string"
          },
          "ensName": {
            "type": "string",
            "nullable": true
          },
          "permissions": {
            "$ref": "#/components/schemas/Permissions"
          }
        },
        "required": ["address"]
      },
...
      "Logout": {
        "type": "string",
        "enum": ["ok"]
      },
...
      "Comment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "replyToId": {
            "type": "string",
            "nullable": true
          },
          "proposalSlug": {
            "type": "string"
          },
          "authorAddress": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "createdAt": {
            "type": "integer"
          },
          "updatedAt": {
            "type": "integer",
            "nullable": true
          }
        },
        "required": ["id", "proposalSlug", "authorAddress", "content", "createdAt"]
      },
      "NewComment": {
        "type": "object",
        "properties": {
          "replyToId": {
            "type": "string",
            "nullable": true
          },
          "content": {
            "type": "string"
          }
        },
        "required": ["content"]
      },
...
    },
    "securitySchemes": {
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "decent-session"
      }
    }
  },
  "security": [
    {
      "CookieAuth": []
    }
  ],
...
}